dev-requirements.txt
google.cloud
google.api_core